e101/pyproject.toml
[project]
name = "agent_clinic_e101"
version = "0.1.0"
description = "Agent Clinic E101"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "functions-framework>=3.10.1",
    "google-adk[eval]>=1.27.1",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
]
